MICS-10338 Standard segment builder fix
When creating a user query segment using the standard segment builder
without adding any items to the query, the newly created segment query
is not editable in navigator (segment page => edit => edit query gives
an error).

This was due to an error in SegmentBuilderPreview since the code
was trying to access the where.expressions fields of a parsed json
that did not contain them.

This commit fixes this, preventing from accessing data that is not
defined, in order to be able to access the edit query part of the
segment's edit form.