MICS-10249 Export segment stats fix
When going to a segment page, and clicking on the "Export stats"
button, an excel file is created and downloaded for the user.

This file contains several tabs. One of them is called "Audience
Segment Add Del". However, the data is missing in that tab. In
order to avoid displaying a tab without data, this commit removes
that tab from the generated excel file.