MICS-11011 Display bug of the tabs on settings
This bug is induced from antd version bump to 4.16.13: class names have
changed and overrides made in `navigatormenu.less` are not effective
anymore.

Chosen solution is then to remove these overrides in order to use
default antd style and avoid this kind of issue in the future when antd
version are upgraded.