MICS-11515 Fix missing compartment id on timeline
When displaying live activities, we do not send the compartment
id to the backend. As the compartment is not sent, the backend
returns live activities based on the datamart default compartment.

This means that we can't see live activities on specific compartment.

To fix this, we just need to send the compartment id with the request.