MICS-11753 Fix Query Tool Page when no multiple datamarts
With MICS-11672, we fixed the fact that the datamart
selector was not working on the query tool page.
Since this fix, the datamart selector is always
displayed, even when the organisation has only
one datamart.

As we don't want to display it in those cases,
we use a default datamart when the organisation
has only one datamart.