MICS-12271 Query tool redirect fix
When going to a link for the query tool with a datamartId in the
query string, the user gets the datamart selector page. In this
case, the user shouldn't get the datamart selector page but the
query tool page directly instead.

This commit fixes this, allowing the query string in the url to
be the only source of truth when getting the selected datamart.