MICS-13141 Interface for Identity Providers
This commit introduces two new interfaces that are used to display
identity provider resources in navigator.
The first one modifies the view in
Settings -> Organisation -> Profile in order to add an Authentication
section that displays the identity provider used by the organisation,
or the default mediarithmics authentication if the organisation does
not have an identity provider.
The second one is a new interface in
Settings -> Organisation -> Identity Providers that is only displayed
for communities. It first displays a table with the identity providers
of the community. It then displays a second table with those identity
providers linked with their associated organisations.