MICS-13539 User and roles modifications
This commit modifies the way users and roles are displayed in
navigator, expecially when there is no user in an organisation.
A text is changed in the user form.
A tooltip is added on "filter on user".
Users in the menu is renamed to Users & Roles.
And finally, after creating a user, a redirection to the role form
is done in order to add a role to the user.