MICS-13741 Users / Roles fixes
This commit adds various fixes concerning users and roles in
navigator.
In the UserRoleForm, when the user input is disabled, it is displayed
with the same style as the organisation input (also disabled).
The way to find the most parent organisation in the UserContainer
is changed so that it searches it within the bounds of the organisation
call for the user.
The Filter on User input width is increased.