MICS-14186 link-on-userpoint
This PR changes the component for the display of results in JSON
in the query tool from a plain <pre></pre> tag to the Ace editor.

This allows the result view to be more easily readable (thanks to
code folding provided by ace). Furthermore, the mode and theme
of the editor are custom, allowing us to display links on user
point ids in the JSON result.

Upon clicking the link, the user is transported to the monitoring tool
in a new tab. The links are displayed when the user point ID is queried,
or another identifier (accounts{id}, agents{id} or emails{id})