MICS-14119 Cohort lookalike fix
The total number of cohorts that could be selected in the
calibration graph depended on the source segment. The computed
cohorts' list was based on the result of the query with the source
segment id, and not the generic query of cohorts in the datamart.
This commit fixes that in order to have the same number of selectable
cohorts in one datamart.