MICS-14726 Feed Card fix
When a feed is active and has a booting feed session, the displayed
feed card is similar to the one for an active feed with a live
feed session.

This commit changes the feed card so that the feed card for an
active feed with a booting feed session is similar to a feed card
for a published feed.

Moreover, this commit modifies the call to feed report in order
to display upserts and deletes in the feed card for the last 24
hours effectively and not since yesterday at 00:00.