MICS-14708 Similarity index cohort calibration
This commit replaces the min_overlap by a similarity_index in the
calibration graph for cohort lookalikes in navigator.
The similarity_index is computed as
overlap / (#(seed_segment) / #(datamart)) with an overlap between 0 and 1.
Similarity indexes in the x axis of the graph are displayed between
its maximum value (which represents the point in which the embedded
cohorts add 10% to the original segment population) and 1, by 0.1
steps.