MICS-14680 Feed cumulative chart
This commit adds a new graph to feed stats. It a cumulative graph
of processed and success upserts minus deletes (for identifiers
or user points), for the last feed session.

If the last feed session started less than 24 hours ago, the
cumulative amount per hour for the last 24 hours is shown. Otherwise,
the graph is shown per day (with a limit of 30 days).