MICS-14782 Fix cumulative feed stats chart
In the cumulative feed chart, when the last feed session was created
less than 24 hours ago, the x axis date values by hour were not
calculated. It is due to the way getAllDates computes its data.
This commit fixes that. Moreover, the from and to in this case
are modified accordingly in FeedCumulativeChart, in order to display
the graph correctly.