MICS-14804 Feed stats tab rework
This commit modifies the feed stats tab in the feed modal. A date
range picker is used, which impacts a new table and a new graph.

The stats are displayed for user points or user identifiers depending
on the feed.

The table displays the number of success upserts, the number of
success deletes and the complement of the percentage of no eligible
identifiers (this one computed for user points) during the given
period of time.

The graph displays the number of success upserts and success deletes
during the given period of time (the time unit being days or hours,
depending on the length of the period of time). This graph is a bar
chart, upserts displayed with positive numbers and deletes with
negative ones.

The previous feed cumulative chart that was in this stats tab is
moved to the troubleshooting tab.